update on shifting on api and splitting connectormanager
