+++ moonshot/src/api/api_environment_variables.py
@@ -1,0 +1,17 @@
+from moonshot.src.configs.env_variables import EnvironmentVars
+
+
+# ------------------------------------------------------------------------------
+# Environment Variables APIs
+# ------------------------------------------------------------------------------
+def api_set_environment_variables(env_vars: dict) -> None:
+    """
+    Sets the environment variables for the current session.
+
+    Args:
+        env_vars (dict): A dictionary containing the environment variables to set.
+
+    Returns:
+        None
+    """
+    EnvironmentVars.load_env(env_vars)
