+++ moonshot/src/utils/import_modules.py
@@ -1,4 +1,5 @@
 import importlib.util
+import inspect
 from importlib.machinery import ModuleSpec
 from types import ModuleType
 from typing import Union
@@ -56,3 +57,38 @@ def import_module_from_spec(module_spec: ModuleSpec) -> Union[ModuleType, None]:
     module = importlib.util.module_from_spec(module_spec)
     module_spec.loader.exec_module(module)
     return module
+
+
+def get_instance(id: str, filepath: str):
+    """
+    A helper method to get an instance of a class from a module
+
+    Args:
+        id (str): The name of the module and the class to be instantiated
+        filepath (str): The path to the module file
+
+    Returns:
+        Any: An instance of the class if found, else None
+    """
+    # Create the module specification
+    module_spec = create_module_spec(
+        id,
+        filepath,
+    )
+
+    # Check if the module specification exists
+    if module_spec:
+        # Import the module
+        module = import_module_from_spec(module_spec)
+
+        # Iterate through the attributes of the module
+        for attr in dir(module):
+            # Get the attribute object
+            obj = getattr(module, attr)
+
+            # Check if the attribute is a class and has the same module name as the id
+            if inspect.isclass(obj) and obj.__module__ == id:
+                return obj
+
+    # Return None if no instance of the class is found
+    return None
